Some more refactorin & breaking
